Verifier/requirements.txt
# backend
fastapi
uvicorn[standard]
python-dotenv

# database
pandas
